# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ccShield backend",
    "description": "Bilibili live-room moderator tool — backend API.",
    "version": "2.0.4"
  },
  "paths": {
...
    "/api/rooms/resolve": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Resolve a B站 room short id or real id to canonical info",
        "description": "Translate ``input`` (real or short id) via B站 HTTP client.\n\nReturns 404 when :meth:`BilibiliClient.resolve_room_id` cannot\nresolve the id (B站 returned a payload without a ``room_id``).",
        "operationId": "resolve_room_route_api_rooms_resolve_get",
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Input"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveRoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/start": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Connect to a live room",
        "description": "Build a :class:`RoomSession`, connect, install a bridge singleton.\n\nFailure path: 400 if ``session.connect`` returns ``False``. The\nsingleton bridge is NOT installed in that case — a failed start\nmust not leave a half-wired bridge behind.",
        "operationId": "start_room_route_api_rooms_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveRoomResponse": {
        "properties": {
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "short_id": {
            "type": "integer",
            "title": "Short Id",
            "default": 0
          },
          "uid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uid"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "uname": {
            "type": "string",
            "title": "Uname",
            "default": ""
          },
          "live_status": {
            "type": "integer",
            "title": "Live Status",
            "default": 0
          },
          "is_short_id": {
            "type": "boolean",
            "title": "Is Short Id",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "room_id"
        ],
        "title": "ResolveRoomResponse",
        "description": "Response body for ``GET /api/rooms/resolve``.\n\n``extra=\"ignore\"`` because B站's payload can carry extra fields that\nare useful for the UI (description, area_id, ...) but we only promise\nthe typed surface here."
      },
...
      "StartRoomRequest": {
        "properties": {
          "room_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Room Id",
            "description": "Real or short room id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "room_id"
        ],
        "title": "StartRoomRequest",
        "description": "Request body for ``POST /api/rooms/start``."
      },
      "StartRoomResponse": {
        "properties": {
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "role": {
            "type": "string",
            "enum": [
              "anchor",
              "admin",
              "viewer",
              "unknown"
            ],
            "title": "Role",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "room_id"
        ],
        "title": "StartRoomResponse",
        "description": "Response body for ``POST /api/rooms/start`` on success."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}